feat: separate 'upgrade', 'containerize'
